pyproject.toml
[project]
name = "nutrition-rag-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.129.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-groq>=1.1.2",
    "pandas>=3.0.0",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "uvicorn>=0.41.0",
]
